Upload home-end recipe (#5595)
* Upload home-end recipe

This package makes HOME/END keys smartly cycle to the beginning/end of a line, the beginning/end of the window, the beginning/end of the buffer, and back to POINT. With a prefix argument, behaves as functions `beginning-of-buffer’/`end-of-buffer’.

A first keypress moves POINT to the beginning/end of a line, or if already there, to the beginning/end of the window, or if already there, to the beginning/end of the buffer. Subsequent keypresses cycle through those operations until returning POINT to its start position. Invoking a PREFIX ARGUMENT prior to the keypress moves POINT consistent with PREFIX ARG M-x beginning-of-buffer / end-of-buffer.

Note that this is a total re-write of an identically-named package that had been bundled into debian’s `emacs-goodies-el’. In 2018, that package was marked for ‘retirement’ due to seeming “to need substantial maintainance and have no upstream home”[1]. The original version had been Copyright 1996 Kai Grossjohann and Toby Speight, and Copyright 2002-2011 Toby Speight, both under GPL3.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=759721#13

* Remove brackets from optional elements

* test

* Change repo name

* Remove 'files' spec